refactor helpers names in perception/ folder
