+++ droidlet/perception/robot/handlers/core.py
@@ -14,7 +14,7 @@ if "/opt/ros/kinetic/lib/python2.7/dist-packages" in sys.path:
     sys.path.remove("/opt/ros/kinetic/lib/python2.7/dist-packages")
 import cv2
 
-from droidlet.perception.robot.perception_helpers import get_color_tag
+from droidlet.perception.robot.perception_util import get_color_tag
 from droidlet.lowlevel.locobot.locobot_mover_utils import xyz_pyrobot_to_canonical_coords
 
 class AbstractHandler:
